Prepare work plan (#2)
TL;DR
-----

Lays out the work plan for the project

Details
-------

Includes the plan for the project. The plan starts with an [initial
prompt](./initial-prompt.md) that was used along with the [source code
for the upstream
project](https://github.com/aws-samples/serverless-pdf-chat) to
develop a [more detailed plan](./prompt-plan.md). The prompt plan in
turn was used to develop a [detailed todo list](./todo.md) that is can
track progress on the project.