Adds lambda functions (#9)
TL;DR
-----

Implements the Lambda functions as part of the app Helm chart

Details
-------

Creates resources for the Lambda functions that are part of the
application.  Each function is created using a function template that
makes their configuration consistent. Each function is implemented
with a Docker image pulled from ECR.

We create the ECR repositories and push the images to the appropriate
repository external to the Helm-based workflows. I tried a few options
to include the functions in to the chart, but none of them worked well
enough---the one that worked best made the Helm chart fill too big.

The `Makefile` has a target named `images` to build and deploy the
images. The `charts` target does not depend on it to speed things up,
but it does need to be run before the charts will work.