Releases the frontend chart with Replicated (#22)
TL;DR
-----

Includes the front end chart in the Replicated release

Details
-------

Updates the Replicated release with the `serverless-pdf-chat` chart
with includes the front-end application. There's more debugging to do
and some enhancements I'd like to add, but I wanted to get this PR in
because other charts needed  some work for pulling images from
Replicated.